refactor(worker): modularize entrypoint and dynamic blog post count
- Split src/index.ts into types, config, util, middleware, registry, fleet, github, tasks, presence, and router modules.
- /posts/count lists clankamode.github.io/posts via the GitHub Contents API, counts YYYY-MM-DD-*.html files, derives latest post from sorted names, caches in KV for one hour (blog:posts:count:v1), and includes that key in admin refresh invalidation.
- Merge github-events.ts into github.ts; index re-exports Env and wires the worker to handleFetch.

Made-with: Cursor
